refactor(settings): extract AppearanceSection to sections/
Moves the Appearance tab — theme picker, typography (font size, font
family), and layout (sidebar width, compact mode) — into its own file
under src/components/settings/sections/.

Prop signatures unchanged (theme + onThemeChange passed from the modal).
No behavior changes.
